refactored next university: BMSTU, SPBSTU, SSEU, TLTSU
