+++ src/test/java/com/techstud/sch_parser/domain/BmstuParserTest.java
@@ -3,8 +3,11 @@ package com.techstud.sch_parser.domain;
 import com.techstud.sch_parser.domain.impl.BmstuParser;
 import com.techstud.sch_parser.domain.impl.SseuParser;
 import com.techstud.sch_parser.model.Schedule;
+import com.techstud.sch_parser.model.api.response.bmstu.BmstuApiResponse;
 import com.techstud.sch_parser.model.kafka.request.ParsingTask;
 import com.techstud.sch_parser.service.MappingService;
+import com.techstud.sch_parser.service.MappingServiceRef;
+import com.techstud.sch_parser.service.impl.BmstuServiceImpl;
 import com.techstud.sch_parser.service.impl.MappingServiceImpl;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.DefaultHttpClient;
@@ -17,19 +20,19 @@ import org.junit.jupiter.api.Test;
 public class BmstuParserTest {
 
     private CloseableHttpClient closeableHttpClient;
-    private MappingService mappingService;
+    private MappingServiceRef<BmstuApiResponse> mappingServiceRef;
 
     @BeforeEach
     public void setUp() {
         closeableHttpClient = new DefaultHttpClient();
-        mappingService = new MappingServiceImpl();
+        mappingServiceRef = new BmstuServiceImpl();
     }
 
     @Test
     public void parseScheduleTest() throws Exception {
         ParsingTask parsingTask = new ParsingTask();
         parsingTask.setGroupId("815c1ebd-bc7e-11ee-b32d-df9b99f124c0");
-        Parser parser = new BmstuParser(closeableHttpClient, mappingService);
+        Parser parser = new BmstuParser(mappingServiceRef);
         Schedule schedule = parser.parseSchedule(parsingTask);
         System.out.println(schedule.toString());
         Assertions.assertNotNull(schedule.toString());
@@ -40,7 +43,7 @@ public class BmstuParserTest {
     public void parseScheduleTestAnother() throws Exception {
         ParsingTask parsingTask = new ParsingTask();
         parsingTask.setGroupId("815c1ebd-bc7e-11ee-b32d-df9b99f124c0");
-        Parser parser = new BmstuParser(closeableHttpClient, mappingService);
+        Parser parser = new BmstuParser(mappingServiceRef);
         Schedule schedule = parser.parseSchedule(parsingTask);
         System.out.println(schedule.toString());
         Assertions.assertNotNull(schedule.toString());
