+++ src/test/java/com/techstud/sch_parser/domain/SpbstuParserTest.java
@@ -5,12 +5,16 @@ import com.techstud.sch_parser.domain.impl.SpbstuParser;
 import com.techstud.sch_parser.model.Schedule;
 import com.techstud.sch_parser.model.kafka.request.ParsingTask;
 import com.techstud.sch_parser.service.MappingService;
+import com.techstud.sch_parser.service.MappingServiceRef;
 import com.techstud.sch_parser.service.impl.MappingServiceImpl;
+import com.techstud.sch_parser.service.impl.SpbstuServiceImpl;
 import lombok.extern.log4j.Log4j2;
+import org.jsoup.nodes.Document;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -20,18 +24,18 @@ import static org.assertj.core.api.Assertions.assertThat;
 @Disabled
 public class SpbstuParserTest {
 
-    private MappingService mappingService;
+    private MappingServiceRef<List<Document>> mappingServiceRef;
 
     @BeforeEach
     public void setUp() {
-        mappingService = new MappingServiceImpl();
+        mappingServiceRef = new SpbstuServiceImpl();
     }
 
     @Test
     public void testReturnSomethingDataByGroupId() throws Exception {
         ParsingTask parsingTask = new ParsingTask();
         parsingTask.setGroupId("41067");
-        Parser underTest = new SpbstuParser(mappingService);
+        Parser underTest = new SpbstuParser(mappingServiceRef);
 
         Optional<Schedule> result = Optional.ofNullable(
                 underTest.parseSchedule(parsingTask));
